+++ alarms-micro/src/main/java/org/opennms/horizon/alarmservice/db/api/AlarmRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlarmRepository extends JpaRepository<Alarm, Integer> {
+public interface AlarmRepository extends JpaRepository<Alarm, Long> {
     @Transactional(Transactional.TxType.REQUIRED)
     public Alarm findByReductionKey(String reductionKey);
 }
